+++ README.md
@@ -6,6 +6,6 @@ For example, you want to see cost of 20 EUR against USD. In this case, EUR is ta
 
 All rates are relevant for this moment.
 
-Moreover, you also can look archive rates. Choose currency pair > choose needed dates > press "Show archive rates" button.
+Moreover, you also can look archive rates. Choose currency pair > choose needed dates > Press "Show archive rates" button.
 
 All data is fetched from [European Central Bank](https://www.ecb.europa.eu/stats/policy_and_exchange_rates/euro_reference_exchange_rates/html/index.en.html). API is provided by [Frankfurter](https://www.frankfurter.app/docs/).
