+++ index.html
@@ -11,6 +11,7 @@
   <body>
     <header>
       <h1 class="header">AccuRate</h1>
+      <img src="resources/logo.png">
     </header>
 
     <main>
@@ -60,6 +61,6 @@
       <ul class="archive-rates-list"></ul>
     </div>
 
-    <script type="module" src="src/app.js"></script>
+    <script type="module" src="src/main.js"></script>
   </body>
 </html>
