+++ src/data/generatearchiverates.js
@@ -1,4 +1,7 @@
-import { targetCoin, baseCoin, fromDate, toDate, archiveRatesList } from '../components/elements.js';
+import { generateArchiveFail } from '../views/generate-archive-fail.js';
+import { generateArchiveSuccess } from '../views/generate-archive-success.js';
+import { targetCoin, baseCoin, fromDate, toDate } from '../views/elements.js';
+
 
 // 5. Shows the exchange rate in previous time periods
 function generateArchiveRates() {
@@ -11,37 +14,9 @@ function generateArchiveRates() {
       return response.json();
     })
     // Success fetching scenario
-    .then(function (data) {
-      // if the page already contains some archive data, it will be deleted
-      archiveRatesList.innerHTML = '';
-    
-      // checks if there is some data about rates > add title 'Archive rates' on the page
-      if (data.rates != '') {
-        const archiveListTitle = document.createElement('h2');
-        archiveListTitle.textContent = 'Archive rates';
-        archiveRatesList.appendChild(archiveListTitle);
-      }
-    
-      // checks if both dates are not empty
-      if ((fromDate.value && toDate.value) === "") {
-        archiveRatesList.innerHTML = 'Please, select time period';
-      }
-      
-      // add data about archive rates on the page
-      for (const day in data.rates) {
-        const archiveElement = document.createElement('li');
-        archiveElement.textContent = `by the date of ${day}, the exchange rate of ${targetCoin.value} was ${Object.values(data.rates[day])} ${baseCoin.value}`;
-        
-        archiveRatesList.appendChild(archiveElement);
-      }
-    })
+    .then(generateArchiveSuccess)
     // Failed fetching scenario
-    .catch(function () {
-      // if the page already contains some archive data, it will be deleted
-      archiveRatesList.innerHTML = '';
-  
-      archiveRatesList.innerHTML = 'No information for this time period or for this currency pair';
-    })
+    .catch(generateArchiveFail)
 }
 
 export { generateArchiveRates }
